+++ app/src/main/kotlin/io/treehouses/remote/DerivedSSHConsole.kt
@@ -28,7 +28,7 @@ open class DerivedSSHConsole: BaseSSHConsole() {
     private var keyRepeatHandler = Handler()
     private val currentPromptHelper: PromptHelper?
         get() {
-            val view = adapter!!.currentTerminalView ?: return null
+            val view = currentTerminalView ?: return null
             return view.bridge.promptHelper
         }
 
@@ -39,93 +39,9 @@ open class DerivedSSHConsole: BaseSSHConsole() {
         }
     }
 
-    inner class TerminalPagerAdapter : PagerAdapter() {
-        override fun getCount(): Int {
-            return if (bound != null) bound!!.bridges.size
-            else 0
-        }
-
-        override fun instantiateItem(container: ViewGroup, position: Int): Any {
-            if (bound == null || bound!!.bridges.size <= position) Log.w(TAG, "Activity not bound when creating TerminalView.")
-            val bridge = bound!!.bridges[position]
-            bridge.promptHelper!!.setHandler(promptHandler)
-
-            // inflate each terminal view
-            val view = layoutInflater.inflate(R.layout.item_terminal, container, false) as RelativeLayout
-
-            // set the terminal name overlay text
-            val terminalNameOverlay = view.findViewById<TextView>(R.id.terminal_name_overlay)
-            terminalNameOverlay.text = bridge.host!!.nickname
-
-            // and add our terminal view control, using index to place behind overlay
-            val terminal = TerminalView(container.context, bridge, bind.pager)
-            terminal.id = R.id.terminal_view
-            view.addView(terminal, 0)
-
-            // Tag the view with its bridge so it can be retrieved later.
-            view.tag = bridge
-            container.addView(view)
-            terminalNameOverlay.startAnimation(fadeOutDelayed)
-            return view
-        }
-
-        override fun destroyItem(container: ViewGroup, position: Int, `object`: Any) {
-            val view = `object` as View
-            container.removeView(view)
-        }
-
-        override fun getItemPosition(`object`: Any): Int {
-            if (bound == null) return POSITION_NONE
-            val view = `object` as View
-            val terminal: TerminalView = view.findViewById(R.id.terminal_view)
-            val host = terminal.bridge.host
-            var itemIndex = POSITION_NONE
-            var i = 0
-            for (bridge in bound!!.bridges) {
-                if (bridge.host == host) {
-                    itemIndex = i
-                    break
-                }
-                i++
-            }
-            return itemIndex
-        }
-
-        fun getBridgeAtPosition(position: Int): TerminalBridge? {
-            if (bound == null) return null
-            val bridges = bound!!.bridges
-            return if (position < 0 || position >= bridges.size) null
-            else bridges[position]
-        }
-
-        override fun notifyDataSetChanged() {
-            super.notifyDataSetChanged()
-            if (tabs != null) {
-                toolbar!!.visibility = if (this.count > 1) View.VISIBLE else View.GONE
-                tabs!!.setTabsFromPagerAdapter(this)
-            }
-        }
-
-        override fun isViewFromObject(view: View, `object`: Any): Boolean {
-            return view === `object`
-        }
-
-        override fun getPageTitle(position: Int): CharSequence? {
-            val bridge = getBridgeAtPosition(position) ?: return "???"
-            return "Treehouses Remote: " + bridge.host!!.nickname
-        }
-
-        val currentTerminalView: TerminalView?
-            get() {
-                val currentView = bind.pager.findViewWithTag<View>(getBridgeAtPosition(bind.pager.currentItem))
-                        ?: return null
-                return currentView.findViewById<View>(R.id.terminal_view) as TerminalView
-            }
-    }
-
     protected fun setUrlItemListener() {
         urlScan!!.setOnMenuItemClickListener {
-            val terminalView = adapter!!.currentTerminalView
+            val terminalView = currentTerminalView
             val urls = terminalView!!.bridge.scanForURLs()
             val urlDialog = Dialog(this@DerivedSSHConsole)
             urlDialog.setTitle("Scan for URLs")
@@ -147,7 +63,13 @@ open class DerivedSSHConsole: BaseSSHConsole() {
                         onTerminalChanged()
                     }
                 })
-        adapter = TerminalPagerAdapter()
+        adapter = TerminalPagerAdapter(bound, promptHandler, layoutInflater, bind, fadeOutDelayed)
+        adapter!!.setTerminalPager(object : TerminalPager {
+            override fun handleData() {
+                handleData()
+            }
+
+        })
         bind.pager.adapter = adapter
         if (tabs != null) setupTabLayoutWithViewPager()
         bind.pager.setOnClickListener { showEmulatedKeys(true) }
@@ -239,7 +161,7 @@ open class DerivedSSHConsole: BaseSSHConsole() {
 
     protected fun addKeyboardListeners() {
         bind.keyboard.buttonKeyboard.setOnClickListener {
-            val terminal = adapter!!.currentTerminalView ?: return@setOnClickListener
+            val terminal = currentTerminalView ?: return@setOnClickListener
             val inputMethodManager = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
             inputMethodManager.toggleSoftInputFromWindow(terminal.applicationWindowToken,
                     InputMethodManager.SHOW_FORCED, 0)
@@ -305,6 +227,13 @@ open class DerivedSSHConsole: BaseSSHConsole() {
             bind.consolePassword.setText("")
             updatePromptVisible()
             true
+        }
+    }
+
+    private fun handleData() {
+        if (tabs != null) {
+            toolbar!!.visibility = if (adapter?.count!! > 1) View.VISIBLE else View.GONE
+            tabs!!.setTabsFromPagerAdapter(adapter)
         }
     }
 }
