+++ app/src/main/kotlin/io/treehouses/remote/TerminalPager.kt
@@ -1,0 +1,5 @@
+package io.treehouses.remote
+
+interface TerminalPager {
+    fun handleData()
+}
